+++ src/multimedia/audio/qsoundeffect.h
@@ -60,7 +60,7 @@ class Q_MULTIMEDIA_EXPORT QSoundEffect : public QObject
     Q_PROPERTY(QUrl source READ source WRITE setSource NOTIFY sourceChanged)
     Q_PROPERTY(int loops READ loopCount WRITE setLoopCount NOTIFY loopCountChanged)
     Q_PROPERTY(int loopsRemaining READ loopsRemaining NOTIFY loopsRemainingChanged)
-    Q_PROPERTY(qreal volume READ volume WRITE setVolume NOTIFY volumeChanged)
+    Q_PROPERTY(float volume READ volume WRITE setVolume NOTIFY volumeChanged)
     Q_PROPERTY(bool muted READ isMuted WRITE setMuted NOTIFY mutedChanged)
     Q_PROPERTY(bool playing READ isPlaying NOTIFY playingChanged)
     Q_PROPERTY(Status status READ status NOTIFY statusChanged)
@@ -94,8 +94,8 @@ public:
     int loopsRemaining() const;
     void setLoopCount(int loopCount);
 
-    qreal volume() const;
-    void setVolume(qreal volume);
+    float volume() const;
+    void setVolume(float volume);
 
     bool isMuted() const;
     void setMuted(bool muted);
